+++ src/mock/products.js
@@ -11,6 +11,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 1,
@@ -19,6 +21,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 2,
@@ -27,6 +31,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 3,
@@ -35,6 +41,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 4,
@@ -43,6 +51,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 5,
@@ -51,6 +61,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 6,
@@ -59,6 +71,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 7,
@@ -67,6 +81,8 @@ export const productList = [
 		image: EX1,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 8,
@@ -75,6 +91,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 9,
@@ -83,6 +101,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 10,
@@ -91,6 +111,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 11,
@@ -99,6 +121,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 12,
@@ -107,6 +131,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 13,
@@ -115,6 +141,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 14,
@@ -123,6 +151,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 15,
@@ -131,6 +161,8 @@ export const productList = [
 		image: EX2,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 16,
@@ -139,6 +171,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 17,
@@ -147,6 +181,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 18,
@@ -155,6 +191,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 19,
@@ -163,6 +201,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 20,
@@ -171,6 +211,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 21,
@@ -179,6 +221,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 22,
@@ -187,6 +231,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 23,
@@ -195,6 +241,8 @@ export const productList = [
 		image: EX3,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 24,
@@ -203,6 +251,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 25,
@@ -211,6 +261,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 26,
@@ -219,6 +271,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 27,
@@ -227,6 +281,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 28,
@@ -235,6 +291,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 29,
@@ -243,6 +301,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 30,
@@ -251,6 +311,8 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 	{
 		id: 31,
@@ -259,5 +321,7 @@ export const productList = [
 		image: EX4,
 		location: "서울시 강남구 역삼2동",
 		price: 50000000,
+		introduction:
+			"개성과 다양성을 상징하는 아디다스 슈퍼스타가 틀을 깨는 새로운 버전으로 돌아왔습니다. 과거에서 얻은 영감을 기반으로 모던한 스트리트웨어에 변화를 주도록 재구성된 아디폼 SST 부츠를 만나보세요. 지속 가능한 소재로 제작된 부츠 컷 디자인이 미래를 향해 진화하는 대담한 룩을 선사하고, 아이코닉한 쉘토가 슈퍼스타의 역사를 재조명합니다. 부드러운 폼 소재를 사용해 편안하고 기분 좋은 착화감을 제공합니다. 한정된 자원을 절약하고 플라스틱 폐기물을 줄이기 위해 천연 원료와 재생 소재로 제작된 친환경 제품입니다.",
 	},
 ];
